+++ interpreter/interpreter.go
@@ -3,7 +3,7 @@ package interpreter
 import (
 	"fmt"
 
-	"github.com/beeceej/iGo/pkg/parse"
+	"github.com/beeceej/iGo/parse"
 )
 
 // Interpreter houses the function references and input history
@@ -23,7 +23,6 @@ func (i *Interpreter) Interpret(text string) {
 	}
 	i.History = append(i.History, text)
 	t := i.classify(text)
-	// spew.Dump(t)
 	for _, tv := range t {
 		switch v := tv.(type) {
 		case *parse.Function:
@@ -44,9 +43,7 @@ func (i *Interpreter) classify(text string) []parse.Classifier {
 		Raw: text,
 	}
 	p.Parse()
-
 	fns := p.Functions
-	fmt.Println(len(fns))
 	if len(fns) == 0 {
 		t = []parse.Classifier{&parse.Expression{Raw: text}}
 	}
